backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
google-generativeai==0.8.3
groq==0.11.0
langchain==0.3.7
langchain-google-genai==2.0.4
langchain-groq==0.2.1
langchain-core==0.3.15
chromadb==0.5.15
pypdf==5.1.0
python-multipart==0.0.12
httpx==0.27.2
